refactor(studio): slice engine.ts into routing/routability/interpolate/dwell modules
