+++ .cortex/skills/build-routing-solution/openrouteservice_app/services/ors_control_app/server/studio/engine.ts
@@ -1,680 +1,38 @@
+// Studio synthetic telemetry orchestrator. The pure helpers
+// (POI loading, routing, interpolation, dwell emission, fleet build, freight
+// offers) live in engine/*; this file only hosts the day-by-day generateTelemetry
+// async generator that ties them together and emits SSE events to jobs.ts.
+
 import {
-  GenerationConfig, DwellConfig, haversineKm, lognormalSample,
-  calculateHeading, addGpsJitter, createRng, rngInt, rngFloat, uuid,
-  resolveVehicleType, VehicleType,
+  GenerationConfig, DwellConfig, createRng, rngInt, resolveVehicleType, uuid,
+  haversineKm,
 } from './profiles.js';
 import { log } from '../diagnostics.js';
+
+import type {
+  POI, RouteGeometry, TelemetryPoint, TripRecord, GenerationEvent,
+  GenerationProgress, SnowSqlFn, VehicleLifecycle,
+} from './engine/types.js';
+
+import { buildFleet } from './engine/fleet.js';
+import { loadPOIs } from './engine/routability.js';
+import {
+  fetchRoute, fetchDetourRoute, pickDestination,
+  pickNearestRoutableNeighbor, pickDetourWaypoint, probeRoutability,
+} from './engine/routing.js';
+import { interpolateRoute } from './engine/interpolate.js';
+import { emitDwell, emitLongIdleDwell } from './engine/dwell.js';
 
 export type {
   TelemetryPoint, TripRecord, GenerationEvent, POI, RouteGeometry,
   FleetMember, GenerationProgress, RouteFetchResult, FreightOffer,
 } from './engine/types.js';
-import type { POI, RouteGeometry, FleetMember, GenerationProgress, RouteFetchResult, FreightOffer, TelemetryPoint, TripRecord, GenerationEvent, SnowSqlFn } from './engine/types.js';
+
+// Re-exports kept for backwards-compatibility with jobs.ts and tests.
 export { generateFreightOffers } from './engine/freight.js';
-
-type VehicleState = 'MOVING' | 'DWELL_ORIGIN' | 'DWELL_DESTINATION' | 'DWELL_REST' | 'DWELL_RECHARGE' | 'IDLE' | 'OVERNIGHT';
-
-interface VehicleLifecycle {
-  vehicle: FleetMember;
-  lat: number;
-  lng: number;
-  currentTime: Date;
-  state: VehicleState;
-  location_id: string | null;
-  location_type: string | null;
-  dailyDrivingMin: number;
-  minSinceBreak: number;
-  tripSeq: number;
-  odometerKm: number;
-  pointIndex: number;
-}
-
-const UNROUTABLE_PATTERNS: RegExp[] = [
-  /Could not find routable point/i,
-  /code['":\s]+2010/i,
-  /point .* not found/i,
-  /coordinate \d+:\s*-?\d+(\.\d+)?\s+-?\d+(\.\d+)?/i,
-];
-
-// Look up ISO-2 country codes for the active region from FLEET_INTELLIGENCE.CORE.REGION_REGISTRY.
-// When the column is non-empty, loadPOIs filters POIs to those countries (eliminates border-bbox
-// leakage). When NULL/empty, no country filter is applied and the job relies on the snap-distance
-// filter + probeRoutability for safety. Returns null on lookup failure (logged WARN, non-fatal).
-async function fetchRegionCountryCodes(region: string, snowSql: SnowSqlFn): Promise<string[] | null> {
-  if (!region) return null;
-  const safe = region.replace(/'/g, "''");
-  try {
-    const rows = await snowSql(
-      `SELECT COUNTRY_CODES FROM FLEET_INTELLIGENCE.CORE.REGION_REGISTRY WHERE REGION_NAME = '${safe}' LIMIT 1`,
-      'FLEET_INTELLIGENCE', 'CORE',
-    );
-    const raw = rows?.[0]?.COUNTRY_CODES;
-    if (raw == null) return null;
-    const arr = Array.isArray(raw) ? raw : (typeof raw === 'string' ? JSON.parse(raw) : null);
-    if (!Array.isArray(arr) || arr.length === 0) return null;
-    return arr.map((c: unknown) => String(c).trim()).filter(Boolean);
-  } catch (e: any) {
-    log('WARN', 'Studio', `REGION_REGISTRY country lookup failed (continuing without country filter): ${e.message?.slice(0, 200)}`, {
-      detail: { region },
-    });
-    return null;
-  }
-}
-
-// Per-profile snap-distance threshold (metres) used by filterRoutablePois.
-// Driving graphs are dense, so a snap > 300 m almost always means the point is
-// off the active country graph (e.g. across a national border). Cycling/foot
-// graphs are sparser and need a wider radius.
-const SNAP_THRESHOLD_M_BY_PROFILE: Record<string, number> = {
-  'driving-car': 300,
-  'driving-hgv': 300,
-  'cycling-regular': 2000,
-  'cycling-electric': 2000,
-  'cycling-mountain': 2000,
-  'cycling-road': 2000,
-  'foot-walking': 2000,
-  'foot-hiking': 2000,
-};
-
-function snapThresholdForProfile(profile: string): number {
-  return SNAP_THRESHOLD_M_BY_PROFILE[profile] ?? 2000;
-}
-
-function isUnroutableError(msg: string): boolean {
-  return UNROUTABLE_PATTERNS.some(p => p.test(msg));
-}
-
-async function filterRoutablePois(
-  pois: POI[],
-  profile: string,
-  region: string,
-  bbox: { min_lat: number; max_lat: number; min_lng: number; max_lng: number },
-  snowSql: SnowSqlFn,
-  onProgressLog?: (msg: string) => void,
-): Promise<POI[]> {
-  if (pois.length === 0) return pois;
-
-  const centerLat = (bbox.min_lat + bbox.max_lat) / 2;
-  const centerLng = (bbox.min_lng + bbox.max_lng) / 2;
-  const sourcesArr = `ARRAY_CONSTRUCT(ARRAY_CONSTRUCT(${centerLng}, ${centerLat}))`;
-  const profileEsc = profile.replace(/'/g, "''");
-  const regionEsc = region.replace(/'/g, "''");
-
-  const BATCH_SIZE = 1000;
-  const SNAP_THRESHOLD_M = snapThresholdForProfile(profile);
-  const reachable = new Array<boolean>(pois.length).fill(false);
-  let droppedNullDuration = 0;
-  let droppedFarSnap = 0;
-
-  for (let i = 0; i < pois.length; i += BATCH_SIZE) {
-    const batch = pois.slice(i, i + BATCH_SIZE);
-    const destsArr = 'ARRAY_CONSTRUCT(' +
-      batch.map(p => `ARRAY_CONSTRUCT(${p.lng}, ${p.lat})`).join(',') +
-      ')';
-    const sql = `
-      SELECT TO_VARCHAR(M:durations[0]) AS DURATIONS,
-             TO_VARCHAR(M:destinations) AS DESTINATIONS
-      FROM (
-        SELECT OPENROUTESERVICE_APP.CORE.MATRIX_TABULAR(
-          '${profileEsc}',
-          ${sourcesArr},
-          ${destsArr},
-          '${regionEsc}'
-        ) AS M
-      )
-    `;
-    try {
-      const rows = await snowSql(sql);
-      const rawDur = rows?.[0]?.DURATIONS;
-      const rawDest = rows?.[0]?.DESTINATIONS;
-      if (!rawDur) {
-        log('WARN', 'Studio', `POI filter batch ${i}-${i + batch.length}: empty result, keeping batch`);
-        for (let j = 0; j < batch.length; j++) reachable[i + j] = true;
-        continue;
-      }
-      const durations = JSON.parse(typeof rawDur === 'string' ? rawDur : String(rawDur));
-      const destinations = rawDest ? JSON.parse(typeof rawDest === 'string' ? rawDest : String(rawDest)) : [];
-      if (!Array.isArray(durations)) {
-        log('WARN', 'Studio', `POI filter batch ${i}: non-array durations, keeping batch`);
-        for (let j = 0; j < batch.length; j++) reachable[i + j] = true;
-        continue;
-      }
-      for (let j = 0; j < batch.length; j++) {
-        const d = durations[j];
-        if (d == null || !Number.isFinite(Number(d))) {
-          droppedNullDuration++;
-          continue;
-        }
-        const dest = Array.isArray(destinations) ? destinations[j] : null;
-        // Treat a null destination object as not routable: ORS could not snap the POI to any
-        // road in the active graph. Older code kept these because snap was undefined.
-        if (dest == null) {
-          droppedNullDuration++;
-          continue;
-        }
-        const snap = dest?.snapped_distance;
-        if (snap == null || !Number.isFinite(Number(snap)) || Number(snap) > SNAP_THRESHOLD_M) {
-          droppedFarSnap++;
-          continue;
-        }
-        reachable[i + j] = true;
-      }
-    } catch (e: any) {
-      log('WARN', 'Studio', `POI filter batch ${i} failed (non-fatal): ${e.message?.slice(0, 200)}`);
-      for (let j = 0; j < batch.length; j++) reachable[i + j] = true;
-    }
-  }
-
-  const filtered = pois.filter((_p, i) => reachable[i]);
-  const dropped = pois.length - filtered.length;
-  log('INFO', 'Studio', `POI routability filter: ${filtered.length}/${pois.length} routable`, {
-    detail: { dropped, droppedNullDuration, droppedFarSnap, profile, region, source: [centerLng, centerLat], snapThresholdM: SNAP_THRESHOLD_M },
-  });
-
-  if (filtered.length < Math.max(50, Math.floor(pois.length * 0.5))) {
-    const msg = `POI filter dropped too many (${dropped}/${pois.length}); falling back to unfiltered list (probable bbox-centroid mismatch with graph)`;
-    log('WARN', 'Studio', msg);
-    onProgressLog?.(`POI filter: ${filtered.length}/${pois.length} routable - too aggressive, using unfiltered list`);
-    return pois;
-  }
-
-  onProgressLog?.(`POI filter: ${filtered.length}/${pois.length} routable (dropped ${droppedNullDuration} unreachable, ${droppedFarSnap} far-snap)`);
-  return filtered;
-}
-
-export async function loadPOIs(
-  config: GenerationConfig,
-  snowSql: SnowSqlFn,
-  onLog?: (msg: string) => void,
-): Promise<POI[]> {
-  const { bbox } = config;
-  const cats = config.poi_categories || ['restaurant', 'bar', 'hotel', 'corporate_or_business_office'];
-  const catFilter = cats.map(c => `'${c}'`).join(',');
-  const countryCodes = await fetchRegionCountryCodes(config.region, snowSql);
-  const countryFilter = countryCodes && countryCodes.length
-    ? `
-      AND p.ADDRESSES[0]:country::STRING IN (${countryCodes.map(c => `'${c.replace(/'/g, "''")}'`).join(',')})`
-    : '';
-  const sql = `
-    WITH region_boundary AS (
-      SELECT BOUNDARY
-      FROM OPENROUTESERVICE_APP.CORE.REGION_CATALOG rc
-      WHERE rc.BOUNDARY IS NOT NULL
-        AND (UPPER(rc.LOOKUP_NAME) = UPPER('${config.region.replace(/'/g, "''")}')
-             OR UPPER(rc.REGION_KEY) = UPPER('${config.region.replace(/'/g, "''")}'))
-      ORDER BY COALESCE(rc.BOUNDARY_AREA_KM2, 1e15) ASC
-      LIMIT 1
-    )
-    SELECT p.ID AS LOCATION_ID, p.NAMES::VARIANT:primary AS NAME,
-           p.BASIC_CATEGORY AS CATEGORY,
-           ST_Y(p.GEOMETRY) AS LAT, ST_X(p.GEOMETRY) AS LNG
-    FROM OVERTURE_MAPS__PLACES.CARTO.PLACE p
-      LEFT JOIN region_boundary rb ON TRUE
-    WHERE ST_Y(p.GEOMETRY) BETWEEN ${bbox.min_lat} AND ${bbox.max_lat}
-      AND ST_X(p.GEOMETRY) BETWEEN ${bbox.min_lng} AND ${bbox.max_lng}
-      AND p.BASIC_CATEGORY IN (${catFilter})${countryFilter}
-      AND COALESCE(ST_INTERSECTS(p.GEOMETRY, rb.BOUNDARY), TRUE)
-    LIMIT 5000`;
-  log('INFO', 'Studio', `Loading POIs from Overture Maps`, {
-    detail: { categories: cats, bbox, mode: config.mode, region: config.region, countryCodes, sql: sql.trim().replace(/\s+/g, ' ') },
-  });
-  try {
-    const rows = await snowSql(sql, 'OVERTURE_MAPS__PLACES', 'CARTO');
-    if (rows.length > 0) {
-      const pois = rows.map((r: any) => ({
-        location_id: r.LOCATION_ID || uuid(Math.random),
-        name: r.NAME || 'Unknown',
-        location_type: mapCategoryToType(r.CATEGORY || '', config.mode),
-        lat: Number(r.LAT),
-        lng: Number(r.LNG),
-        category: r.CATEGORY || '',
-      }));
-      const catCounts: Record<string, number> = {};
-      const typeCounts: Record<string, number> = {};
-      for (const p of pois) {
-        catCounts[p.category] = (catCounts[p.category] || 0) + 1;
-        typeCounts[p.location_type] = (typeCounts[p.location_type] || 0) + 1;
-      }
-      log('INFO', 'Studio', `Loaded ${pois.length} POIs from Overture Maps`, {
-        detail: { source: 'overture', categories: catCounts, types: typeCounts },
-      });
-      const sanitized = await filterRoutablePois(pois, config.ors_profile, config.region, bbox, snowSql, onLog);
-      return sanitized;
-    }
-    log('ERROR', 'Studio', `Overture Maps returned 0 POIs for bbox`, {
-      detail: { bbox, categories: cats },
-    });
-    throw new Error(
-      `No POIs found in Overture Maps for region bbox ` +
-      `[${bbox.min_lat},${bbox.min_lng} to ${bbox.max_lat},${bbox.max_lng}] ` +
-      `with categories [${cats.join(', ')}]. Expand the bbox or change categories.`
-    );
-  } catch (e: any) {
-    if (e.message?.startsWith('No POIs found')) throw e;
-    log('ERROR', 'Studio', `Overture Maps query failed`, {
-      detail: { error: e.message?.slice(0, 200), bbox, categories: cats },
-    });
-    throw new Error(
-      `Cannot load POIs: Overture Maps is not accessible. ` +
-      `Ensure the OVERTURE_MAPS__PLACES share is mounted. Error: ${e.message?.slice(0, 200)}`
-    );
-  }
-}
-
-// Pre-flight probe: pick N random ordered pairs from the (post-filter) POI list and
-// confirm they actually route on the active graph. Catches any remaining POI/graph
-// mismatch before vehicles are generated. For multi-country regions (no country filter)
-// this is the only safety net, so we keep it conservative.
-export async function probeRoutability(
-  pois: POI[],
-  profile: string,
-  region: string,
-  snowSql: SnowSqlFn,
-  opts?: { sampleSize?: number; minSuccess?: number; rng?: () => number },
-): Promise<{ ok: boolean; success: number; total: number; failures: Array<{ origin: [number, number]; dest: [number, number]; reason: string }> }> {
-  const sampleSize = opts?.sampleSize ?? 5;
-  const minSuccess = opts?.minSuccess ?? 3;
-  const rng = opts?.rng ?? Math.random;
-  if (pois.length < 2) {
-    return { ok: false, success: 0, total: 0, failures: [{ origin: [0, 0], dest: [0, 0], reason: 'fewer than 2 POIs available' }] };
-  }
-  const failures: Array<{ origin: [number, number]; dest: [number, number]; reason: string }> = [];
-  let success = 0;
-  for (let i = 0; i < sampleSize; i++) {
-    const a = pois[Math.floor(rng() * pois.length)];
-    let b = pois[Math.floor(rng() * pois.length)];
-    let guard = 0;
-    while (b.location_id === a.location_id && guard++ < 10) b = pois[Math.floor(rng() * pois.length)];
-    const result = await fetchRoute(a.lat, a.lng, b.lat, b.lng, profile, region, snowSql);
-    if (result && result !== 'UNROUTABLE') {
-      success++;
-    } else {
-      failures.push({ origin: [a.lat, a.lng], dest: [b.lat, b.lng], reason: result === 'UNROUTABLE' ? 'UNROUTABLE' : 'hard_fail' });
-    }
-  }
-  return { ok: success >= minSuccess, success, total: sampleSize, failures };
-}
-
-function mapCategoryToType(category: string, mode: string): string {
-  if (mode === 'food_delivery') {
-    if (['restaurant', 'fast_food_restaurant', 'cafe', 'bakery', 'pizzaria', 'casual_eatery', 'coffee_shop', 'sandwich_shop', 'chicken_restaurant'].includes(category)) return 'RESTAURANT';
-    return 'ADDRESS';
-  }
-  if (mode === 'trucking') {
-    if (['warehouse', 'storage_facility', 'b2b_transportation_and_storage_service', 'industrial_facility_or_service'].includes(category)) return 'WAREHOUSE';
-    if (['gas_station', 'parking', 'transportation_location', 'ground_transport_facility_or_service'].includes(category)) return 'REST_STOP';
-    return 'DESTINATION';
-  }
-  return 'LOCATION';
-}
-
 export { buildFleet } from './engine/fleet.js';
-import { buildFleet } from './engine/fleet.js';
-
-
-async function fetchRoute(
-  originLat: number, originLng: number,
-  destLat: number, destLng: number,
-  profile: string,
-  region: string,
-  snowSql: SnowSqlFn,
-): Promise<RouteFetchResult> {
-  const sql = `
-    SELECT TO_VARCHAR(ST_ASGEOJSON(GEOJSON)) AS GEO_STR, DISTANCE, DURATION
-    FROM TABLE(OPENROUTESERVICE_APP.CORE.DIRECTIONS(
-      '${profile}',
-      ARRAY_CONSTRUCT(${originLng},${originLat}),
-      ARRAY_CONSTRUCT(${destLng},${destLat}),
-      '${region.replace(/'/g, "''")}'
-    ))`;
-  try {
-    const rows = await snowSql(sql);
-    if (!rows.length) {
-      return 'UNROUTABLE';
-    }
-    const dist = rows[0].DISTANCE;
-    const dur = rows[0].DURATION;
-    const geo = typeof rows[0].GEO_STR === 'string' ? JSON.parse(rows[0].GEO_STR) : rows[0].GEO_STR;
-    const coords: [number, number][] = geo?.coordinates || [];
-    if (coords.length < 2 || dist == null || dur == null) {
-      return 'UNROUTABLE';
-    }
-    return {
-      coordinates: coords.map(c => [c[1], c[0]]),
-      distance_m: Number(dist) || 0,
-      duration_sec: Number(dur) || 0,
-    };
-  } catch (e: any) {
-    const msg = String(e?.message || '');
-    if (isUnroutableError(msg)) {
-      return 'UNROUTABLE';
-    }
-    log('WARN', 'Studio', `Route fetch failed: ${msg.slice(0, 300)}`, {
-      detail: { origin: [originLat, originLng], dest: [destLat, destLng], profile },
-    });
-    return null;
-  }
-}
-
-async function fetchDetourRoute(
-  originLat: number, originLng: number,
-  waypointLat: number, waypointLng: number,
-  destLat: number, destLng: number,
-  profile: string,
-  region: string,
-  snowSql: SnowSqlFn,
-): Promise<RouteFetchResult> {
-  const coordsJson = JSON.stringify({
-    coordinates: [
-      [originLng, originLat],
-      [waypointLng, waypointLat],
-      [destLng, destLat],
-    ],
-  }).replace(/'/g, "''");
-  const sql = `
-    SELECT TO_VARCHAR(ST_ASGEOJSON(GEOJSON)) AS GEO_STR, DISTANCE, DURATION
-    FROM TABLE(OPENROUTESERVICE_APP.CORE.DIRECTIONS(
-      '${profile}',
-      PARSE_JSON('${coordsJson}')::VARIANT,
-      '${region.replace(/'/g, "''")}'
-    ))`;
-  try {
-    const rows = await snowSql(sql);
-    if (!rows.length) return 'UNROUTABLE';
-    const dist = rows[0].DISTANCE;
-    const dur = rows[0].DURATION;
-    const geo = typeof rows[0].GEO_STR === 'string' ? JSON.parse(rows[0].GEO_STR) : rows[0].GEO_STR;
-    const coords: [number, number][] = geo?.coordinates || [];
-    if (coords.length < 2 || dist == null || dur == null) return 'UNROUTABLE';
-    return {
-      coordinates: coords.map(c => [c[1], c[0]]),
-      distance_m: Number(dist) || 0,
-      duration_sec: Number(dur) || 0,
-    };
-  } catch (e: any) {
-    const msg = String(e?.message || '');
-    if (isUnroutableError(msg)) {
-      return 'UNROUTABLE';
-    }
-    log('WARN', 'Studio', `Detour route fetch failed: ${msg.slice(0, 300)}`, {
-      detail: { profile },
-    });
-    return null;
-  }
-}
-
-function pickDestination(
-  origin: POI, pois: POI[], config: GenerationConfig, rng: () => number,
-): POI {
-  const destPois = pois.filter(p => p.location_id !== origin.location_id);
-  if (destPois.length === 0) return origin;
-
-  const { short_pct, short_max_km, medium_pct, medium_max_km } = config.distance_distribution;
-  const r = rng();
-  let maxKm: number;
-  if (r < short_pct) maxKm = short_max_km;
-  else if (r < short_pct + medium_pct) maxKm = medium_max_km;
-  else maxKm = 999;
-
-  const nearby = destPois.filter(p => haversineKm(origin.lat, origin.lng, p.lat, p.lng) <= maxKm);
-  const pool = nearby.length > 0 ? nearby : destPois;
-  return pool[Math.floor(rng() * pool.length)];
-}
-
-function pickNearestRoutableNeighbor(
-  origin: POI, pois: POI[], rng: () => number,
-): POI | null {
-  const NEIGHBOR_RADIUS_KM = 10;
-  const candidates = pois.filter(p =>
-    p.location_id !== origin.location_id &&
-    haversineKm(origin.lat, origin.lng, p.lat, p.lng) <= NEIGHBOR_RADIUS_KM
-  );
-  if (candidates.length === 0) return null;
-  return candidates[Math.floor(rng() * candidates.length)];
-}
-
-function pickDetourWaypoint(
-  origin: POI, dest: POI, pois: POI[], rng: () => number,
-): POI | null {
-  const midLat = (origin.lat + dest.lat) / 2;
-  const midLng = (origin.lng + dest.lng) / 2;
-  const directDist = haversineKm(origin.lat, origin.lng, dest.lat, dest.lng);
-  const maxOffset = directDist * 0.5;
-  const candidates = pois.filter(p =>
-    p.location_id !== origin.location_id &&
-    p.location_id !== dest.location_id &&
-    haversineKm(midLat, midLng, p.lat, p.lng) <= maxOffset
-  );
-  if (candidates.length === 0) return null;
-  return candidates[Math.floor(rng() * candidates.length)];
-}
-
-function interpolateRoute(
-  route: RouteGeometry, config: GenerationConfig,
-  lifecycle: VehicleLifecycle, tripId: string,
-  destPoi: POI, rng: () => number, isDetour: boolean,
-): TelemetryPoint[] {
-  const points: TelemetryPoint[] = [];
-  const coords = route.coordinates;
-  if (coords.length < 2) return points;
-  const vt = resolveVehicleType(config);
-
-  const segments: number[] = [0];
-  let totalDist = 0;
-  for (let i = 1; i < coords.length; i++) {
-    totalDist += haversineKm(coords[i - 1][0], coords[i - 1][1], coords[i][0], coords[i][1]);
-    segments.push(totalDist);
-  }
-  if (totalDist === 0) return points;
-
-  const durationSec = route.duration_sec || (totalDist / (lifecycle.vehicle.base_speed_kmh / 3.6));
-  const pingMean = config.telemetry.ping_interval_moving.mean_sec;
-  const pingStd = config.telemetry.ping_interval_moving.std_sec;
-  const jitterCfg = config.telemetry.gps_jitter;
-  const postedSpeeds = config.routing.posted_speeds;
-  const defaultSpeed = postedSpeeds.default || 30;
-  const speedingThreshold = 1.08;
-
-  let elapsed = 0;
-  const vehicle = lifecycle.vehicle;
-  const hosMaxDriveMin = config.breaks?.max_daily_driving_hours
-    ? config.breaks.max_daily_driving_hours * 60 : Infinity;
-
-  while (elapsed < durationSec) {
-    if (vt === 'hgv' && lifecycle.dailyDrivingMin >= hosMaxDriveMin) {
-      break;
-    }
-
-    const progress = Math.min(elapsed / durationSec, 1);
-    const distAtProgress = progress * totalDist;
-
-    let segIdx = 0;
-    for (let i = 1; i < segments.length; i++) {
-      if (segments[i] >= distAtProgress) { segIdx = i - 1; break; }
-      segIdx = i - 1;
-    }
-    segIdx = Math.min(segIdx, coords.length - 2);
-    const segStart = segments[segIdx];
-    const segEnd = segments[segIdx + 1] || segStart + 0.001;
-    const segFrac = (distAtProgress - segStart) / (segEnd - segStart);
-
-    const lat = coords[segIdx][0] + segFrac * (coords[segIdx + 1][0] - coords[segIdx][0]);
-    const lng = coords[segIdx][1] + segFrac * (coords[segIdx + 1][1] - coords[segIdx][1]);
-
-    const speedFactor = 1 + (rng() - 0.5) * vehicle.speed_variance * 2;
-    let speedKmh = vehicle.base_speed_kmh * speedFactor;
-    if (progress < 0.1 || progress > 0.9) speedKmh *= 0.7;
-    speedKmh = Math.max(5, Math.min(speedKmh, 130));
-    const postedSpeed = defaultSpeed;
-    const isSpeeding = rng() < vehicle.speeding_prob && speedKmh > postedSpeed * speedingThreshold;
-    if (isSpeeding) speedKmh = postedSpeed * rngFloat(rng, 1.1, 1.25);
-
-    const isHosViolation = vt === 'hgv' &&
-      !!config.breaks?.max_daily_driving_hours &&
-      lifecycle.dailyDrivingMin > config.breaks.max_daily_driving_hours * 60 &&
-      rng() < vehicle.hos_violation_prob;
-
-    const heading = calculateHeading(lat, lng,
-      coords[Math.min(segIdx + 1, coords.length - 1)][0],
-      coords[Math.min(segIdx + 1, coords.length - 1)][1]);
-
-    const jitterM = rng() < jitterCfg.multipath_probability
-      ? rngFloat(rng, 50, jitterCfg.multipath_max_m)
-      : rngFloat(rng, 2, jitterCfg.typical_m);
-    const [jLat, jLng] = addGpsJitter(lat, lng, jitterM, rng);
-
-    const ts = new Date(lifecycle.currentTime.getTime() + elapsed * 1000);
-
-    let batteryPct: number | null = null;
-    if (vt === 'ebike' && config.battery) {
-      const kmTraveled = distAtProgress;
-      const drain = kmTraveled * config.battery.drain_per_km;
-      batteryPct = Math.max(0, vehicle.battery_pct - drain);
-    }
-
-    lifecycle.odometerKm += (speedKmh / 3600) * pingMean;
-
-    points.push({
-      telemetry_id: uuid(rng),
-      region: config.region,
-      vehicle_type: vt,
-      vehicle_id: vehicle.vehicle_id,
-      trip_id: tripId,
-      ts,
-      latitude: jLat,
-      longitude: jLng,
-      speed_kmh: Math.round(speedKmh * 10) / 10,
-      heading_deg: Math.round(heading * 10) / 10,
-      posted_speed_kmh: postedSpeed,
-      status: 'MOVING',
-      is_speeding: isSpeeding,
-      is_hos_violation: isHosViolation,
-      is_detour: isDetour,
-      gps_accuracy_m: jitterM,
-      location_id: null,
-      location_type: null,
-      ors_profile: config.ors_profile,
-      battery_pct: batteryPct,
-      odometer_km: Math.round(lifecycle.odometerKm * 100) / 100,
-      point_index: lifecycle.pointIndex++,
-    });
-
-    const interval = Math.max(5, pingMean + (rng() - 0.5) * 2 * pingStd);
-    elapsed += interval;
-  }
-
-  lifecycle.lat = coords[coords.length - 1][0];
-  lifecycle.lng = coords[coords.length - 1][1];
-  lifecycle.currentTime = new Date(lifecycle.currentTime.getTime() + durationSec * 1000);
-  lifecycle.dailyDrivingMin += durationSec / 60;
-  lifecycle.minSinceBreak += durationSec / 60;
-
-  if (vt === 'ebike' && config.battery) {
-    const kmTraveled = totalDist;
-    vehicle.battery_pct = Math.max(0, vehicle.battery_pct - kmTraveled * config.battery.drain_per_km);
-  }
-
-  return points;
-}
-
-function emitDwell(
-  lifecycle: VehicleLifecycle, config: GenerationConfig, tripId: string | null,
-  dwellConfig: DwellConfig, status: VehicleState, poi: POI | null,
-  rng: () => number,
-): TelemetryPoint[] {
-  const points: TelemetryPoint[] = [];
-  const vt = resolveVehicleType(config);
-  let dwellMin = lognormalSample(dwellConfig.median_min, dwellConfig.sigma, dwellConfig.max_min, rng);
-  if (dwellConfig.long_wait_probability && rng() < dwellConfig.long_wait_probability) {
-    dwellMin = rngFloat(rng, dwellConfig.max_min, dwellConfig.max_min * 2);
-  }
-  const dwellSec = dwellMin * 60;
-  const pingMin = config.telemetry.ping_interval_dwell.min_sec;
-  const pingMax = config.telemetry.ping_interval_dwell.max_sec;
-
-  let elapsed = 0;
-  while (elapsed < dwellSec) {
-    const [jLat, jLng] = addGpsJitter(lifecycle.lat, lifecycle.lng, 3, rng);
-    const ts = new Date(lifecycle.currentTime.getTime() + elapsed * 1000);
-    points.push({
-      telemetry_id: uuid(rng),
-      region: config.region,
-      vehicle_type: vt,
-      vehicle_id: lifecycle.vehicle.vehicle_id,
-      trip_id: tripId,
-      ts,
-      latitude: jLat,
-      longitude: jLng,
-      speed_kmh: 0,
-      heading_deg: 0,
-      posted_speed_kmh: 0,
-      status,
-      is_speeding: false,
-      is_hos_violation: false,
-      is_detour: false,
-      gps_accuracy_m: 3,
-      location_id: poi?.location_id || null,
-      location_type: poi?.location_type || null,
-      ors_profile: config.ors_profile,
-      battery_pct: lifecycle.vehicle.battery_pct > 0 ? lifecycle.vehicle.battery_pct : null,
-      odometer_km: Math.round(lifecycle.odometerKm * 100) / 100,
-      point_index: lifecycle.pointIndex++,
-    });
-    elapsed += rngFloat(rng, pingMin, pingMax);
-  }
-  lifecycle.currentTime = new Date(lifecycle.currentTime.getTime() + dwellSec * 1000);
-  lifecycle.state = status;
-  lifecycle.location_id = poi?.location_id || null;
-  lifecycle.location_type = poi?.location_type || null;
-  return points;
-}
-
-// Emit a single multi-day IDLE dwell session for "ghost trailers". Pings are
-// sparse (default 5-15 min) so the row volume stays modest even over a 7-day
-// window. The dwell sessionizer (CONDITIONAL_CHANGE_EVENT on STATUS) will roll
-// these contiguous IDLE pings into one DT_DWELL_SESSIONS row.
-function emitLongIdleDwell(
-  lifecycle: VehicleLifecycle, config: GenerationConfig, poi: POI | null,
-  durationSec: number, pingMinSec: number, pingMaxSec: number, rng: () => number,
-): TelemetryPoint[] {
-  const points: TelemetryPoint[] = [];
-  const vt = resolveVehicleType(config);
-  let elapsed = 0;
-  while (elapsed < durationSec) {
-    const [jLat, jLng] = addGpsJitter(lifecycle.lat, lifecycle.lng, 2, rng);
-    const ts = new Date(lifecycle.currentTime.getTime() + elapsed * 1000);
-    points.push({
-      telemetry_id: uuid(rng),
-      region: config.region,
-      vehicle_type: vt,
-      vehicle_id: lifecycle.vehicle.vehicle_id,
-      trip_id: null,
-      ts,
-      latitude: jLat,
-      longitude: jLng,
-      speed_kmh: 0,
-      heading_deg: 0,
-      posted_speed_kmh: 0,
-      status: 'IDLE',
-      is_speeding: false,
-      is_hos_violation: false,
-      is_detour: false,
-      gps_accuracy_m: 3,
-      location_id: poi?.location_id || null,
-      location_type: poi?.location_type || null,
-      ors_profile: config.ors_profile,
-      battery_pct: lifecycle.vehicle.battery_pct > 0 ? lifecycle.vehicle.battery_pct : null,
-      odometer_km: Math.round(lifecycle.odometerKm * 100) / 100,
-      point_index: lifecycle.pointIndex++,
-    });
-    elapsed += rngFloat(rng, pingMinSec, pingMaxSec);
-  }
-  lifecycle.currentTime = new Date(lifecycle.currentTime.getTime() + durationSec * 1000);
-  lifecycle.state = 'IDLE';
-  lifecycle.location_id = poi?.location_id || null;
-  lifecycle.location_type = poi?.location_type || null;
-  return points;
-}
+export { loadPOIs } from './engine/routability.js';
+export { probeRoutability } from './engine/routing.js';
 
 export async function* generateTelemetry(
   config: GenerationConfig,
